refactor: reorg signals in community app
